requirements.txt
lmdb
requests
opencv-python
Shapely==1.5.13
numpy==1.15.1
matplotlib==1.5.1
scipy==0.19.0
plumbum==1.6.2
ipython==6.1.0
Pillow==4.2.1
torch==0.4.0
torchvision==0.2.1
tensorflow-gpu==1.4.0
tornado
asyncio
ImageHash
pyyaml